pyproject.toml
[project]
name = "ner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "evaluate>=0.4.3",
    "mistralai>=1.2.1",
    "openai>=1.53.0",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.5.2",
    "seqeval>=1.2.2",
    "torch>=2.5.1",
    "transformers[torch]>=4.46.1",
]
